+++ docs/reference/workspace-layout.md
@@ -91,6 +91,9 @@ Key paths:
   `host_macos` / `host_effects_macos` transition modules
 - `packages/rsnap-overlay/src/overlay.rs`: current overlay root plus its focused
   runtime/rendering support modules
+- `packages/rsnap-overlay/src/overlay/rendering.rs`: overlay window renderer orchestration; shared
+  selection geometry/cache structs live under `overlay/rendering/selection_geometry.rs`, and
+  renderer phase timing telemetry lives under `overlay/rendering/timing.rs`
 - `packages/rsnap-overlay/src/live_frame_stream_macos.rs`: current macOS live-stream support
 - `packages/rsnap-overlay/src/scroll_capture.rs`: current scroll-capture session entry with
   focused support modules under `scroll_capture/`
